move tentative structure to base for now
